+++ src/com/neilatkinson/gameobject/Updateable.java
@@ -3,6 +3,6 @@ package com.neilatkinson.gameobject;
 public interface Updateable {
 
 	public void update();
-	public void animate(long elapsedTime);
+	public void animate(int elapsedTime);
 
 }
